Refactor: 파일정리, 셀 분리, 뷰 같이 쓰기
